pyproject.toml
[project]
name = "find-broken-desktop"
dynamic = ["version"]
description = "Find desktop entry files (*.desktop) with broken executables."
authors = [{ name = "smac89", email = "noblechuk5[at]web[dot]de" }]
requires-python = ">=3.10"
readme = "README.md"
keywords = ["desktop", "xdg", "executable", "broken"]
license = { text = "GPL" }

dependencies = ["pyxdg==0.28"]

[project.urls]
Repository = "https://github.com/smac89/find-broken-desktop"

[dependency-groups]
dev = ["ruff>=0.14.2"]

[project.scripts]
find-broken-desktop = "fbrokendesktop.__main__:main"

[tool.ruff]
fix = true
line-length = 120

[tool.ruff.lint]
select = [
    # https://docs.astral.sh/ruff/settings/#lint_select
    "E4",
    "E7",
    "E9",
    "F",
    "B",
    "F",
    "I",
]

[tool.ruff.format]
# Enable reformatting of code snippets in docstrings.
docstring-code-format = true

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"
fallback_version = "1.4.0"
write_to = "fbrokendesktop/_version.py"
write_template = "__version__ = '{}'"

[build-system]
requires = ["pdm-backend >= 2.4.0"]
build-backend = "pdm.backend"
